+++ src/main/java/com/ibc/training/fooddelivery/repository/OrderRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Long> {
+public interface OrderRepository extends JpaRepository<Order, Integer> {
 
-    List<Order> findByCustomerId(Long customerId);
+    List<Order> findByCustomerId(Integer customerId);
 
-    List<Order> findByDriverId(Long driverId);
+    List<Order> findByDriverId(Integer driverId);
 }
